Lots of trial and error with Azure. Added some local llama scripts to play with, nothing serious.
$$Important$$
Added diarizeAndTranscribe.py inside of a new directory called PyannoteAndWhisper.

This is a very important file as it takes input wavs of any duration and diarizes and transcribes with time stamps.

I think in next meeting we need to push some things to the master branch as things will be a bit messy on main after this.